requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Flask==2.2.3
Flask-WTF==1.1.1
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1